Add initial SDNS client and server implementations
Add sdns-proxy, which is meant to run on client machines to proxy DNS
queries to either a public DNS server or the internal "SDNS" server (for
lack of a better name).

Add sdns-server, which is meant to run on a server on the local network to
handle secured DNS queries to private resources. Currently only returns
hard-coded mappings, but this is intended to be extended to manage virtual
IPs.

Also include a self-signed SSL certificate for testing.